Moving thrift package into api package
