+++ src/app/test/api/thrift/handler_test.py
@@ -7,7 +7,7 @@ from pika.exceptions import UnroutableError
 from pyrabbit2.http import HTTPError
 
 import beer_garden.bg_utils
-from beer_garden.thrift.handler import BartenderHandler
+from beer_garden.api.thrift.handler import BartenderHandler
 from brewtils.errors import ModelValidationError
